feat: consolidate mixins; simplify parameters
