+++ sqlspec/statement/pipelines/context.py
@@ -8,7 +8,7 @@ from sqlspec.exceptions import RiskLevel
 if TYPE_CHECKING:
     from sqlglot.dialects.dialect import DialectType
 
-    from sqlspec.statement.parameters import ParameterInfo, ParameterNormalizationState
+    from sqlspec.statement.parameters import ParameterInfo, ParameterStyleTransformationState
     from sqlspec.statement.sql import SQLConfig
     from sqlspec.typing import SQLParameterType
 
@@ -99,7 +99,7 @@ class SQLProcessingContext:
     extra_info: dict[str, Any] = field(default_factory=dict)
     """Extra information from parameter processing, including normalization state."""
 
-    parameter_normalization: "Optional[ParameterNormalizationState]" = None
+    parameter_normalization: "Optional[ParameterStyleTransformationState]" = None
     """Single source of truth for parameter normalization tracking."""
 
     @property
