+++ tests/unit/test_statement/test_builder/test_builder_mixins.py
@@ -27,15 +27,10 @@ import pytest
 from sqlglot import exp
 
 from sqlspec.exceptions import SQLBuilderError
-from sqlspec.statement.builder.mixins._aggregate_functions import AggregateFunctionsMixin
-from sqlspec.statement.builder.mixins._from import FromClauseMixin
-from sqlspec.statement.builder.mixins._group_by import GroupByClauseMixin
-from sqlspec.statement.builder.mixins._having import HavingClauseMixin
-from sqlspec.statement.builder.mixins._insert_from_select import InsertFromSelectMixin
-from sqlspec.statement.builder.mixins._insert_values import InsertValuesMixin
-from sqlspec.statement.builder.mixins._join import JoinClauseMixin
-from sqlspec.statement.builder.mixins._limit_offset import LimitOffsetClauseMixin
-from sqlspec.statement.builder.mixins._merge_clauses import (
+from sqlspec.statement.builder.mixins.cte_and_set_ops import SetOperationMixin
+from sqlspec.statement.builder.mixins.insert_operations import InsertFromSelectMixin, InsertValuesMixin
+from sqlspec.statement.builder.mixins.join_operations import JoinClauseMixin
+from sqlspec.statement.builder.mixins.merge_operations import (
     MergeIntoClauseMixin,
     MergeMatchedClauseMixin,
     MergeNotMatchedBySourceClauseMixin,
@@ -43,14 +38,15 @@ from sqlspec.statement.builder.mixins._merge_clauses import (
     MergeOnClauseMixin,
     MergeUsingClauseMixin,
 )
-from sqlspec.statement.builder.mixins._order_by import OrderByClauseMixin
-from sqlspec.statement.builder.mixins._pivot import PivotClauseMixin
-from sqlspec.statement.builder.mixins._returning import ReturningClauseMixin
-from sqlspec.statement.builder.mixins._set_ops import SetOperationMixin
-from sqlspec.statement.builder.mixins._unpivot import UnpivotClauseMixin
-from sqlspec.statement.builder.mixins._update_from import UpdateFromClauseMixin
-from sqlspec.statement.builder.mixins._update_set import UpdateSetClauseMixin
-from sqlspec.statement.builder.mixins._where import WhereClauseMixin
+from sqlspec.statement.builder.mixins.order_limit_operations import (
+    LimitOffsetClauseMixin,
+    OrderByClauseMixin,
+    ReturningClauseMixin,
+)
+from sqlspec.statement.builder.mixins.pivot_operations import PivotClauseMixin, UnpivotClauseMixin
+from sqlspec.statement.builder.mixins.select_operations import SelectClauseMixin
+from sqlspec.statement.builder.mixins.update_operations import UpdateFromClauseMixin, UpdateSetClauseMixin
+from sqlspec.statement.builder.mixins.where_clause import HavingClauseMixin, WhereClauseMixin
 
 if TYPE_CHECKING:
     from sqlglot.dialects.dialect import DialectType
@@ -153,7 +149,7 @@ def test_where_clause_basic(condition: Any, expected_type: type[exp.Expression])
 def test_where_clause_wrong_expression_type() -> None:
     """Test WHERE clause with wrong expression type."""
     builder = WhereTestBuilder(exp.Insert())
-    with pytest.raises(SQLBuilderError, match="Cannot add WHERE clause to unsupported expression type"):
+    with pytest.raises(SQLBuilderError, match="WHERE clause not supported for Insert"):
         builder.where("id = 1")
 
 
@@ -369,7 +365,7 @@ def test_order_by_wrong_expression_type() -> None:
         builder.order_by("name")
 
 
-class FromTestBuilder(MockBuilder, FromClauseMixin):
+class FromTestBuilder(MockBuilder, SelectClauseMixin):
     """Test builder with FROM clause mixin."""
 
     pass
@@ -543,7 +539,7 @@ def test_set_operation_wrong_expression_type() -> None:
         builder1.union(builder2)
 
 
-class GroupByTestBuilder(MockBuilder, GroupByClauseMixin):
+class GroupByTestBuilder(MockBuilder, SelectClauseMixin):
     """Test builder with GROUP BY mixin."""
 
     pass
@@ -944,7 +940,7 @@ def test_unpivot_wrong_expression_type() -> None:
         builder.unpivot(value_column_name="value", name_column_name="name", columns_to_unpivot=["col1"])
 
 
-class AggregateTestBuilder(MockBuilder, AggregateFunctionsMixin):
+class AggregateTestBuilder(MockBuilder, SelectClauseMixin):
     """Test builder with aggregate functions mixin."""
 
     def select(self, expr: Any) -> "AggregateTestBuilder":
